Summary survival curve from aggregated survival data of a meta-analysis

Estimation of the summary survival curve from the survival rates and the numbers of at-risk individuals extracted from studies of a meta-analysis.

msurv(
  study,
  time,
  n.risk,
  surv.rate,
  confidence,
  correctionFlag = TRUE,
  correctionVal = c(0.25, 0.5)
)

Arguments

study

A numeric vector with the numbering of the studies included in the meta-analysis. The numbering of a study is repeated for each survival probabilities extracted from this study.

time

A numeric vector with the time at which the survival probabilities are collected.

n.risk

A numeric vector with the number of at-risk patients in the study for each value of thr time.

surv.rate

A numeric vector with the survival rates collected per study for each value of time.

confidence

A text argument indicating the method to calculate the confidence interval of the summary survival probabilities: "Greenwood" or "MonteCarlo".

correctionFlag

A logical variable which takes into account if user wants the continuity correaction or not (By default TRUE).

correctionVal

A numeric vector for continuity correction, if you don't want to apply correction pass c(0,0).

Value

list

Author

Shubhram Pandey shubhram1992@gmail.com

Examples

attach(exampleData)
result <- msurv(study = Study,
                time = Time,
                 n.risk = NbRisk,
                 surv.rate = Survival,
                 confidence = "Greenwood",
                 correctionFlag = FALSE
                 )
result
#> $verif.data
#>    Sstudy check
#> 1       1     1
#> 2       2     1
#> 3       3     1
#> 4       4     1
#> 5       5     1
#> 6       6     1
#> 7       7     1
#> 8       8     1
#> 9       9     1
#> 10     10     1
#> 11     11     1
#> 12     12     1
#> 13     13     1
#> 14     14     1
#> 15     15     1
#> 16     16     1
#> 17     17     1
#> 18     18     1
#> 19     19     1
#> 20     20     1
#> 21     21     1
#> 22     22     1
#> 23     23     1
#> 24     24     1
#> 25     25     1
#> 26     26     1
#> 27     27     1
#> 
#> $summary.fixed
#>       IndiceTimes PooledSurvivalFE PooledSurvivalICinfFE PooledSurvivalICsupFE
#>  [1,]           1        0.9508837            0.93115618             0.9710293
#>  [2,]           2        0.8542395            0.82202310             0.8877185
#>  [3,]           3        0.7529489            0.71356234             0.7945096
#>  [4,]           4        0.6801444            0.63752795             0.7256097
#>  [5,]           5        0.6081467            0.56352148             0.6563058
#>  [6,]           6        0.5449696            0.49936809             0.5947354
#>  [7,]           9        0.4303415            0.38446782             0.4816886
#>  [8,]          12        0.3452733            0.30011887             0.3972213
#>  [9,]          15        0.2814425            0.23729653             0.3338012
#> [10,]          18        0.2411057            0.19773211             0.2939935
#> [11,]          21        0.2109529            0.16828161             0.2644443
#> [12,]          24        0.1858428            0.14380950             0.2401619
#> [13,]          27        0.1671095            0.12528742             0.2228921
#> [14,]          30        0.1549979            0.11269730             0.2131760
#> [15,]          33        0.1322367            0.08722757             0.2004703
#> [16,]          36        0.1201395            0.07462625             0.1934106
#> [17,]          39        0.1201395            0.07462625             0.1934106
#> [18,]          42        0.1144864            0.06665886             0.1966300
#> [19,]          45        0.1144864            0.06665886             0.1966300
#> [20,]          48        0.1144864            0.06665886             0.1966300
#>       nRisk
#>  [1,]  1806
#>  [2,]  1674
#>  [3,]  1467
#>  [4,]  1265
#>  [5,]  1122
#>  [6,]   974
#>  [7,]   807
#>  [8,]   546
#>  [9,]   368
#> [10,]   252
#> [11,]   187
#> [12,]   142
#> [13,]   100
#> [14,]    65
#> [15,]    37
#> [16,]    29
#> [17,]    12
#> [18,]    12
#> [19,]    10
#> [20,]     9
#> 
#> $median.fixed
#>              2.5%    97.5% 
#> 7.176926 5.937078 8.375936 
#> 
#> $mean.fixed
#>              2.5%    97.5% 
#> 13.46370 11.46318 14.74525 
#> 
#> $heterogeneity
#> [1] 219.6071244   0.7572659   0.0000000
#> 
#> $summary.random
#>       IndiceTimes PooledSurvivalRE PooledSurvivalICinfRE PooledSurvivalICsupRE
#>  [1,]           1       0.94768647            0.91731470             0.9790638
#>  [2,]           2       0.84755282            0.78823478             0.9113348
#>  [3,]           3       0.73182809            0.65606515             0.8163402
#>  [4,]           4       0.65409564            0.56903170             0.7518757
#>  [5,]           5       0.57872161            0.49473929             0.6769600
#>  [6,]           6       0.50184228            0.41034422             0.6137425
#>  [7,]           9       0.38269038            0.29711268             0.4929171
#>  [8,]          12       0.29890154            0.22167479             0.4030324
#>  [9,]          15       0.23629264            0.16877051             0.3308292
#> [10,]          18       0.20208143            0.14274382             0.2860853
#> [11,]          21       0.17082971            0.11580178             0.2520064
#> [12,]          24       0.15126442            0.10079984             0.2269937
#> [13,]          27       0.13757394            0.09019805             0.2098337
#> [14,]          30       0.12300453            0.07677350             0.1970747
#> [15,]          33       0.10536383            0.06049565             0.1835097
#> [16,]          36       0.10074068            0.05720945             0.1773952
#> [17,]          39       0.10074068            0.05720945             0.1773952
#> [18,]          42       0.08499342            0.03811229             0.1895420
#> [19,]          45       0.08499342            0.03811229             0.1895420
#> [20,]          48       0.08499342            0.03811229             0.1895420
#>       nRisk
#>  [1,]  1806
#>  [2,]  1674
#>  [3,]  1467
#>  [4,]  1265
#>  [5,]  1122
#>  [6,]   974
#>  [7,]   807
#>  [8,]   546
#>  [9,]   368
#> [10,]   252
#> [11,]   187
#> [12,]   142
#> [13,]   100
#> [14,]    65
#> [15,]    37
#> [16,]    29
#> [17,]    12
#> [18,]    12
#> [19,]    10
#> [20,]     9
#> 
#> $median.random
#>              2.5%    97.5% 
#> 6.046385 4.815541 8.474948 
#> 
#> $mean.random
#>              2.5%    97.5% 
#> 11.92947  9.13597 14.39214 
#>